Updated minor issues

+++ headers/Utilities.h
@@ -11,7 +11,7 @@ void HandleInput();
 void Exp1();
 void Exp2();
 void ExecuteExp(int bytes, int ways, int expNumber);
-void saveFiles(unsigned int  hitRatio[], int bytes, int ways, int expNumber);
+void saveFiles(double  hitRatio[], int bytes, int ways, int expNumber);
 unsigned int GetAddress(int j, Random& randGen1, Random& randGen2);
 void error(void);
